reworked callbackData logic
in first implementation callbackData was set with key word for operation + title name for example: "REMOVE" + title.
The problem is that max length for setCallbackData() method is 64 bytes, due to lots  of chars in method it wont fit for some animes.
Now instead of title it uses anime id and everything should work fine.